+++ env.example
@@ -1,0 +1,4 @@
+GRAPHQL_SDK_QUERIES_DIRECTORY=./src/graphql
+GRAPHQL_SDK_HASURA_URL=http://localhost:8080/v1/graphql
+GRAPHQL_SDK_HASURA_ADMIN_TOKEN=myadminsecretkey
+GRAPHQL_SDK_OUTPUT=./src/lib/sdk.ts
