+++ src/index.ts
@@ -1,4 +1,3 @@
 // Main entry point for the library
-export { default as codegenConfig } from './codegen';
 export { default as SubscriptionDocLookupPlugin } from './plugins/subscription-doc-lookup';
-export { createCodegenConfig, createSerializableConfig, type GraphQLSDKOptions } from './config-factory';
+export { createCodegenConfig } from './codegen';
